% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{getGenomicVariants}
\alias{getGenomicVariants}
\alias{getGenomicVariants,KatDetect-method}
\title{Retrieve genomic variants from KatDetect object.}
\usage{
getGenomicVariants(x)

\S4method{getGenomicVariants}{KatDetect}(x)
}
\arguments{
\item{x}{(KatDetect): KatDetect object.}
}
\value{
(VRanges): Returns a VRanges with annotated genomic variants.
}
\description{
Retrieve genomic variants from KatDetect object.
}
\examples{
x <- new('KatDetect', genomicVariants = VariantAnnotation::VRanges(), info = list())
getGenomicVariants(x)

}
