% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{getInfo}
\alias{getInfo}
\alias{getInfo,KatDetect-method}
\title{Retrieve model parameters from a KatDetect object.}
\usage{
getInfo(x)

\S4method{getInfo}{KatDetect}(x)
}
\arguments{
\item{x}{(KatDetect): KatDetect object.}
}
\value{
(list): Returns a list with all model parameters used for kataegis detection.
}
\description{
Retrieve model parameters from a KatDetect object.
}
\examples{
kd <- constructKatdetect()
getInfo(kd)

}
