% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{getKataegisFoci}
\alias{getKataegisFoci}
\alias{getKataegisFoci,KatDetect-method}
\title{Retrieve kateagis foci from a KatDetect object.}
\usage{
getKataegisFoci(x)

\S4method{getKataegisFoci}{KatDetect}(x)
}
\arguments{
\item{x}{(KatDetect): KatDetect object.}
}
\value{
(GRanges): Returns a GRanges with annotated kataegis foci.
}
\description{
Retrieve kateagis foci from a KatDetect object.
}
\examples{
kd <- constructKatdetect()
getKataegisFoci(kd)

}
