% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{getSegments}
\alias{getSegments}
\alias{getSegments,KatDetect-method}
\title{Retrieve segments from KatDetect a object.}
\usage{
getSegments(x)

\S4method{getSegments}{KatDetect}(x)
}
\arguments{
\item{x}{(KatDetect): KatDetect object.}
}
\value{
(GRanges): Returns a GRanges with annotated segments.
}
\description{
Retrieve segments from KatDetect a object.
}
\examples{
x <- new('KatDetect', genomicVariants = VariantAnnotation::VRanges(), info = list())
getSegments(x)

}
