% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/katdetectr.R
\docType{package}
\name{katdetectr}
\alias{katdetectr}
\title{Katdetectr: A package for kataegis detection, characterization and visualization}
\description{
The katdetectr package provides three main functions:
detectKataegis(), rainfallPlot() and generateSyntheticData()
}
\details{
See the vignette for more details, a step by step explanation of these function, and the general katdetectr workflow.
}
