% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_detectKataegis.R
\name{detectKataegis}
\alias{detectKataegis}
\title{detectKataegis}
\usage{
detectKataegis(
  genomicVariants,
  minSizeKataegis = 6,
  maxMeanIMD = 1000,
  test.stat = "Exponential",
  penalty = "BIC",
  pen.value = 0,
  minseglen = 2,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1),
  aggregateRecords = FALSE
)
}
\arguments{
\item{genomicVariants}{(\link[VariantAnnotation]{VRanges}, VCF or MAF): VRanges, or path to VCF or MAF file containing genomic variants.}

\item{minSizeKataegis}{(integer): Minimal number of variants required within a segment for classification as a kataegis foci.}

\item{maxMeanIMD}{(integer): Max. mean IMD within a segment for classification as a kataegis foci.}

\item{test.stat}{(character): Distribution that is fitted to the data (Exponential or Empirical). See \link[changepoint]{cpt.meanvar}.}

\item{penalty}{(character): Penalty used to guard against overfitting (BIC or Manual). See \link[changepoint]{cpt.meanvar}.}

\item{pen.value}{(integer): Only needed for manual penalty. See \link[changepoint]{cpt.meanvar}.}

\item{minseglen}{(integer): Min. size of segments (no. of variants).}

\item{BPPARAM}{(\link\href{BiocParallelParam}{BiocParallel}): Can be used for parallelization purposes.}

\item{aggregateRecords}{(logical): Aggregate multiple samples and treat as-if all records originate from a single sample.}
}
\value{
(KatDetect): Returns a KatDetect object including putative kataegis foci.
}
\description{
Detection of kataegis foci using changepoint detection.

Changepoint detection is performed on the intermutation distance (IMD)
of the variants using the changepoint \insertCite{killick2014changepoint}{katdetectr} package.

Note that we recommend using the default parameters for the detection of kataegis.
}
\examples{
syntheticData <- generateSyntheticData()
kd <- detectKataegis(syntheticData)

}
\references{
\insertRef{killick2014changepoint}{katdetectr}
}
\author{
Daan Hazelaar

Job van Riet
}
